requirements.txt
aniso8601==8.0.0
bcrypt==3.1.7
blinker==1.4
boto3==1.12.18
botocore==1.15.18
cffi==1.14.0
click==7.1.1
colorama==0.4.3
cryptography==2.9.2
docutils==0.15.2
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.1
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
loguru==0.4.1
MarkupSafe==1.1.1
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-engineio==3.13.1
python-socketio==4.6.0
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
Werkzeug==1.0.0
win32-setctime==1.0.1
zope.event==4.4
zope.interface==5.1.0
